chatbot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
PyYAML==6.0.2
requests==2.32.3
qdrant-client==1.12.1
pytest==8.3.4

